refactor(ivy): break compiler canonical test into smaller files
