+++ packages/core/test/render3/compiler_canonical/life_cycle_spec.ts
@@ -1,0 +1,107 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {ChangeDetectionStrategy, ChangeDetectorRef, Component, ContentChild, ContentChildren, Directive, HostBinding, HostListener, Injectable, Input, NgModule, OnDestroy, Optional, Pipe, PipeTransform, QueryList, SimpleChanges, TemplateRef, ViewChild, ViewChildren, ViewContainerRef} from '../../../src/core';
+import * as $r3$ from '../../../src/core_render3_private_export';
+import {renderComponent, toHtml} from '../render_util';
+
+/// See: `normative.md`
+describe('lifecycle hooks', () => {
+  let events: string[] = [];
+  let simpleLayout: SimpleLayout;
+
+  type $boolean$ = boolean;
+  type $LifecycleComp$ = LifecycleComp;
+  type $SimpleLayout$ = SimpleLayout;
+
+  beforeEach(() => { events = []; });
+
+  @Component({selector: 'lifecycle-comp', template: ``})
+  class LifecycleComp {
+    @Input('name') nameMin: string;
+
+    ngOnChanges() { events.push('changes' + this.nameMin); }
+
+    ngOnInit() { events.push('init' + this.nameMin); }
+    ngDoCheck() { events.push('check' + this.nameMin); }
+
+    ngAfterContentInit() { events.push('content init' + this.nameMin); }
+    ngAfterContentChecked() { events.push('content check' + this.nameMin); }
+
+    ngAfterViewInit() { events.push('view init' + this.nameMin); }
+    ngAfterViewChecked() { events.push('view check' + this.nameMin); }
+
+    ngOnDestroy() { events.push(this.nameMin); }
+
+    // NORMATIVE
+    static ngComponentDef = $r3$.ɵdefineComponent({
+      type: LifecycleComp,
+      tag: 'lifecycle-comp',
+      factory: function LifecycleComp_Factory() { return new LifecycleComp(); },
+      template: function LifecycleComp_Template(ctx: $LifecycleComp$, cm: $boolean$) {},
+      inputs: {nameMin: 'name'},
+      inputsPropertyName: {nameMin: 'nameMin'},
+      features: [$r3$.ɵNgOnChangesFeature]
+    });
+    // /NORMATIVE
+  }
+
+  @Component({
+    selector: 'simple-layout',
+    template: `
+      <lifecycle-comp [name]="name1"></lifecycle-comp>
+      <lifecycle-comp [name]="name2"></lifecycle-comp>
+    `
+  })
+  class SimpleLayout {
+    name1 = '1';
+    name2 = '2';
+
+    // NORMATIVE
+    static ngComponentDef = $r3$.ɵdefineComponent({
+      type: SimpleLayout,
+      tag: 'simple-layout',
+      factory: function SimpleLayout_Factory() { return simpleLayout = new SimpleLayout(); },
+      template: function SimpleLayout_Template(ctx: $SimpleLayout$, cm: $boolean$) {
+        if (cm) {
+          $r3$.ɵE(0, LifecycleComp);
+          $r3$.ɵe();
+          $r3$.ɵE(2, LifecycleComp);
+          $r3$.ɵe();
+        }
+        $r3$.ɵp(0, 'name', $r3$.ɵb(ctx.name1));
+        $r3$.ɵp(2, 'name', $r3$.ɵb(ctx.name2));
+        LifecycleComp.ngComponentDef.h(1, 0);
+        LifecycleComp.ngComponentDef.h(3, 2);
+        $r3$.ɵr(1, 0);
+        $r3$.ɵr(3, 2);
+      }
+    });
+    // /NORMATIVE
+  }
+
+  it('should gen hooks with a few simple components', () => {
+    expect(toHtml(renderComponent(SimpleLayout)))
+        .toEqual(`<lifecycle-comp></lifecycle-comp><lifecycle-comp></lifecycle-comp>`);
+    expect(events).toEqual([
+      'changes1', 'init1', 'check1', 'changes2', 'init2', 'check2', 'content init1',
+      'content check1', 'content init2', 'content check2', 'view init1', 'view check1',
+      'view init2', 'view check2'
+    ]);
+
+    events = [];
+    simpleLayout.name1 = '-one';
+    simpleLayout.name2 = '-two';
+    $r3$.ɵdetectChanges(simpleLayout);
+    expect(events).toEqual([
+      'changes-one', 'check-one', 'changes-two', 'check-two', 'content check-one',
+      'content check-two', 'view check-one', 'view check-two'
+    ]);
+  });
+
+});
